Web sockets setup

Closes #78

See merge request cse1105/2023-2024/teams/oopp-team-55!97

+++ server/src/main/java/server/service/ExpenseService.java
@@ -12,14 +12,18 @@ import java.util.Optional;
 public class ExpenseService {
 
     private final ExpenseRepository expenseRepository;
+    private final EventService eventService;
 
     @Autowired
-    public ExpenseService(ExpenseRepository expenseRepository) {
+    public ExpenseService(ExpenseRepository expenseRepository, EventService eventService) {
         this.expenseRepository = expenseRepository;
+        this.eventService = eventService;
     }
 
     public Expense saveExpense(Expense expense) {
-        return expenseRepository.saveAndFlush(expense);
+        expense = expenseRepository.saveAndFlush(expense);
+        eventService.updateLastActivity(expense.getEvent().getId());
+        return expense;
     }
 
     public Optional<Expense> getExpenseById(Long id) {
@@ -35,6 +39,12 @@ public class ExpenseService {
     }
 
     public void deleteExpense(Long id) {
+        Optional<Expense> getRes = getExpenseById(id);
+        if (getRes.isEmpty()) {
+            return;
+        }
+        Expense expense = getRes.get();
         expenseRepository.deleteById(id);
+        eventService.updateLastActivity(expense.getEvent().getId());
     }
 }
